+++ x-pack/test/alerting_api_integration/common/lib/get_proxy_server.ts
@@ -4,32 +4,24 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import http from 'http';
 import httpProxy from 'http-proxy';
-import { ToolingLog } from '@kbn/dev-utils';
 
 export const getHttpProxyServer = async (
-  defaultKibanaTargetUrl: string,
+  targetUrl: string,
   kbnTestServerConfig: any,
-  log: ToolingLog
-): Promise<http.Server> => {
-  const proxy = httpProxy.createProxyServer({ secure: false, selfHandleResponse: false });
-
-  const proxyPort = getProxyPort(kbnTestServerConfig);
-  const proxyServer = http.createServer((req: http.IncomingMessage, res: http.ServerResponse) => {
-    const targetUrl = new URL(req.url ?? defaultKibanaTargetUrl);
-
-    if (targetUrl.hostname !== 'some.non.existent.com') {
-      proxy.web(req, res, {
-        target: `${targetUrl.protocol}//${targetUrl.hostname}:${targetUrl.port}`,
-      });
-    } else {
-      res.writeHead(500, { 'Content-Type': 'text/plain' });
-      res.write('error on call some.non.existent.com');
-      res.end();
-    }
+  onProxyResHandler: (proxyRes?: unknown, req?: unknown, res?: unknown) => void
+): Promise<httpProxy> => {
+  const proxyServer = httpProxy.createProxyServer({
+    target: targetUrl,
+    secure: false,
+    selfHandleResponse: false,
   });
 
+  proxyServer.on('proxyRes', (proxyRes: unknown, req: unknown, res: unknown) => {
+    onProxyResHandler(proxyRes, req, res);
+  });
+
+  const proxyPort = getProxyPort(kbnTestServerConfig);
   proxyServer.listen(proxyPort);
 
   return proxyServer;
